UI issue fixed for detail screen
UI issue fixed for detail screen

+++ TSP/Model/BillDetails/SingleBillDetailsModel.swift
@@ -27,7 +27,7 @@ struct SingleBillDetails: Codable {
 
 struct MultipleBillDetails: Codable {
     let status, txnId: String?
-    let amount: Int?
+    let amount: Double?
     let billId, paymentMode, customerName, billerID: String?
     let email: String?
     let noOfInstallment: Int?
